# Improve avatar handling