--- a/GitCommands/Settings/AppSettings.cs
+++ b/GitCommands/Settings/AppSettings.cs
@@ -598,6 +598,7 @@ namespace GitCommands
             set => SetInt("Appearance.AvatarCacheSize", value);
         }
 
+        // Currently not configurable in UI
         // Names from here: https://learn.microsoft.com/en-us/dotnet/api/system.windows.media.brushes?view=windowsdesktop-7.0
         // or #AARRGGBB code
         public static string AvatarAuthorInitialsPalette
@@ -605,6 +606,9 @@ namespace GitCommands
             get => GetString("Appearance.AvatarAuthorInitialsPalette", "SlateGray,RoyalBlue,Purple,OrangeRed,Teal,OliveDrab");
             set => SetString("Appearance.AvatarAuthorInitialsPalette", value);
         }
+
+        // Currently not configurable in UI
+        public static float AvatarAuthorInitialsLuminanceThreshold => GetFloat("AvatarAuthorInitialsLuminanceThreshold", 0.5f);
 
         /// <summary>
         /// Loads a setting with GetString and parses it to an enum