--- a/GitUI/Avatars/FileSystemAvatarCache.cs
+++ b/GitUI/Avatars/FileSystemAvatarCache.cs
@@ -26,10 +26,10 @@ namespace GitUI.Avatars
         /// <inheritdoc />
         public async Task<Image?> GetAvatarAsync(string email, string? name, int imageSize)
         {
-            var cacheDir = AppSettings.AvatarImageCachePath;
-            var path = Path.Combine(cacheDir, $"{email}.{imageSize}px.png");
+            string cacheDir = AppSettings.AvatarImageCachePath;
+            string path = Path.Combine(cacheDir, $"{email}.{imageSize}px.png");
 
-            var image = ReadImage() ?? await _inner.GetAvatarAsync(email, name, imageSize);
+            Image image = ReadImage() ?? await _inner.GetAvatarAsync(email, name, imageSize);
 
             if (image is not null)
             {
@@ -47,12 +47,24 @@ namespace GitUI.Avatars
 
                 try
                 {
-                    using var output = _fileSystem.File.OpenWrite(path);
+                    using Stream output = _fileSystem.File.OpenWrite(path);
                     image.Save(output, ImageFormat.Png);
                 }
                 catch
                 {
-                    // do nothing
+                    // Workaround to avoid the "A generic error occurred in GDI+." exception when saving
+                    // where copying the image in a new one allows the save on disk to be successful...
+                    try
+                    {
+                        using (Bitmap newImage = new(image))
+                        {
+                            using Stream output = _fileSystem.File.OpenWrite(path);
+                            newImage.Save(output, ImageFormat.Png);
+                        }
+                    }
+                    catch
+                    {
+                    }
                 }
             }
 
@@ -62,7 +74,7 @@ namespace GitUI.Avatars
                 {
                     try
                     {
-                        using var stream = _fileSystem.File.OpenRead(path);
+                        using Stream stream = _fileSystem.File.OpenRead(path);
                         return Image.FromStream(stream);
                     }
                     catch