… Plugins/GitUIPluginInterfaces/ISettingsSource.cs
… UnitTests/GitCommands.Tests/Settings/AppSettingsTests.cs
… UnitTests/GitExtUtils.Tests/ColorTransformationTests.cs
… UnitTests/GitUI.Tests/Avatars/InitialsAvatarGeneratorTests.cs