<!-- Please read CONTRIBUTING.md before submitting a pull request -->

Improvements over #10020

Note: a lot easier to review commit by commit.

## Proposed changes

* perf: increase a lot perf to download avatars (ex on a blame display: from 1m30s to 6s)
* perf: fix serialization that was broken for some avatars
* Always obtain the same Author initials image between 2 GitExtensions session
* return more meaningful author initials
* Better set of colors for Author initials background colors
* perf: make that cache never expire for avatar that never changes
* Perf: slight optimization of GravatarProvider
* Perf: prevent requesting the same avatar multiple times
* saner performance default values

## Screenshots <!-- Remove this section if PR does not change UI -->

For choice of colors:

### Before
| Before  |  After |
|---|---|
| ![image](https://user-images.githubusercontent.com/460196/229131809-a53290da-24c6-4c71-bde0-a76fdc2d08eb.png)  | ![image](https://user-images.githubusercontent.com/460196/229132118-f5cb5af7-0074-489b-a0b1-80335dd6da44.png) |

## Test methodology <!-- How did you ensure quality? -->

- Manual
- Some unit tests

## Test environment(s) <!-- Remove any that don't apply -->

- Git Extensions 33.33.33
- Build de0dcb3112d960166620a45a5bf6e49f6918ea3b (Dirty)
- Git 2.38.1.windows.1
- Microsoft Windows NT 10.0.19045.0
- .NET 6.0.14
- DPI 96dpi (no scaling)
- Microsoft.WindowsDesktop.App Versions

## Merge strategy

<!-- Change the following if the merge strategy should be changed:
- Squash merge (maintainer to decide merge message, PR submitter should cleanup commits/messages at PR approval).
- Rebase merge (PR submitter must change the commit message for the last commit).
- Merge commit. (PR submitter to rebase and squash before merges).
- To be decided later.
The maintainer may still request the contributor to squash and rebase, to make sure that merges and commit messages are clarified.
-->

Merge commit

----

:black_nib: I contribute this code under [The Developer Certificate of Origin](../blob/master/contributors.txt).