@vbjay 2023-04-01
No prob. Why I sent it was exactly for the idea of when theming works. The
dark theme can have a gradient of light color backgrounds to show a good
contrast on top of a dark background. Along with text color being good on
said background. Then flip the idea on light backgrounds.

I used chatgpt and thought it would be cool to just copy the original
convo.

On Sat, Apr 1, 2023, 3:58 PM Gerhard Olsson ***@***.***>
wrote:

> ***@***.**** approved this pull request.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/gitextensions/gitextensions/pull/10846#pullrequestreview-1367919288>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AA3WDMN4V2TVKE534BBFR7TW7CCG7ANCNFSM6AAAAAAWOVCCQE>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>