@vbjay 2023-04-03
Just an improvement on my contrasting colors comment

https://gist.github.com/vbjay/cc56f5e67a9165bd022418167b6af323

![image](https://user-images.githubusercontent.com/3629489/229387245-94902b0c-e427-43f0-a180-3a665770773a.png)

Open in linqpad and run.  Play with  thresholds and such to tweak.