@vbjay 2023-04-03
I'd change the getconteastingcolor function use the dynamic thresholds idea. That allows us to tweak the low or high value if needed.  

This is an extension method and we could make it part of GitGui extensions. That way if something else needs contrasting color it is reusable and consistent. 


https://gist.github.com/vbjay/cc56f5e67a9165bd022418167b6af323