@mstv 2023-04-03
(GH supports delta reviews: blue button "Changes since last review". I find this very useful. But GH loses track after too many rebases / force pushes. For the future, please rebase only in order to resolve conflicts or perhaps after a longer inactivity. Thank you.)