@pmiossec 2023-04-04
> (GH supports delta reviews: blue button "Changes since last review". I find this very useful. But GH loses track after too many rebases / force pushes. For the future, please rebase only in order to resolve conflicts or perhaps after a longer inactivity. Thank you.)

I didn't know this option that looks great. I will keep for now on the fixup/rebase as much as possible for a last pre-merge step.