@pmiossec 2023-04-04
> I'd change the getcontrastingcolor function use the dynamic thresholds idea. That allows us to tweak the low or high value if needed.
> 
> This is an extension method and we could make it part of GitGui extensions. That way if something else needs contrasting color it is reusable and consistent.
> 
> https://gist.github.com/vbjay/cc56f5e67a9165bd022418167b6af323
> 
> Could even adjust extension method to have an overload that accepts the thresholds if we wanted to override for some reason.

@vbjay Reading multiple times the code, I absolutely don't understand threshold part. Either it is very clever and complicated and I don't get it or the implementation is bugged. I tend for the 2nd option without being certain: why changing the threshold value based on the luminance and after comparing them whereas a simple comparison seems enough? 

So I have made something simpler that I think is enough...

Have a look, please.