@vbjay 2023-04-04
The idea is if needed you can set <.5 threshold to .2 and >.5 to .8 and it adjusts the points where color flips to white. 


Open script in gist with linkpad.net and play with adjusting threshold values and running.