@pmiossec 2023-04-04
> The idea is if needed you can set <.5 threshold to .2 and >.5 to .8 and it adjusts the points where color flips to white.

That's what my code is doing also, I think.
And playing with linqpad does not help understand the code....
I spent int total more than 20min in the 2 loc below without understanding the added value to change the threshold based on the luminance value **AND AFTER** comparing these values instead of changing the threshold value directly which is way more clearer.

Because, for me it's a lot of complexity for something wrong. Let's see with the 0.2 and 0.8 values you are mentioning, so the code:

```
	// Use a dynamic threshold based on the luminance of the background color.  Play with threshold values when <.5 or not to tweak the flip
	double threshold = (luminance < 0.5) ? 0.2 : 0.8;

	// Use the threshold value to determine whether to use black or white as the foreground color
	Color foregroundColor = (luminance > threshold) ? Color.Black : Color.White;
```

The results:

| | luminance < 0.2 |  0.2 < luminance < 0.5 | 0.5 < luminance < 0.8  | luminance > 0.8  |
|-|-----------------|---|---|---|
|    Background |   Very dark | dark  | light | Very light  |
|Threshold (updated)  | 0.2  |  0.2  |  0.8  | 0.8  |
| Foreground |     `false` => White       | `true` => Black  | `false` => White  | `true` => Black  |
| What we want? |    White on Very dark:👍           | Black on dark:👎  |  White on light:👎 | Black on Very light:👍  |

For me the only good behavior is when the 2 values are the same which is equivalent to my (a lot simpler to understand) code...

Or I misunderstood everything and that's not all the values are expected to be set? In this case, can you explain in details?