@RussKie 2023-04-07
@vbjay has approved, so I think it is.

On Sat, 8 Apr 2023, 8:44 am Gerhard Olsson, ***@***.***>
wrote:

> Ready to merge?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/gitextensions/gitextensions/pull/10846#issuecomment-1500703169>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABBTEXVLX7K4QOO7HYIEVFTXACKFFANCNFSM6AAAAAAWOVCCQE>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>