@vbjay 2023-04-07
Correct.  I was just pushing for a little more control over the threshold and fine tuning if contrast didn't quite work.