@pmiossec 2023-04-08
> Correct. I was just pushing for a little more control over the threshold and fine tuning if contrast didn't quite work.

Yes but except if I missed something, your proposition is not working or not adding value, with much more complexity, as I shown in my list comment. I was waiting for your answer on that.

If you are all OK, I could rebase and squash one or 2 commits to be ready to go...