@vbjay 2023-04-08
Last attempt. Think of it instead of a single slider you have 2. If low you can independently adjust where the threshold if less than .5 or darker colors. This .5 threshold isn't perfect and allowing 2 thresholds lets us independently adjust the thresholds if light or dark.