@vbjay 2023-04-08
The other point. I wanted you to keep it sn extension method. So if and when a color needs to get contrast it is c.GetContrastColor