@pmiossec 2023-04-09
> Last attempt. Think of it instead of a single slider you have 2. If low you can independently adjust where the threshold if less than .5 or darker colors. This .5 threshold isn't perfect and allowing 2 thresholds lets us independently adjust the thresholds if light or dark.

I understand the general idea but I would have preferred an explanation with a real example with working threshold values because except if you found an error in the one I gave, we see that the solution with 2 threshold values is not working and add a lot of complexity.

Also, having only one setting is nearly self-explanatory if someone wants to play with it. Good luck for the explanation and the documentation if we include the 2 threshold settings... I spend quite some time investigating it and I still can't get the added value...(for me it's not working!) 

I will be interested by another point of view on the subject. 

> The other point. I wanted you to keep it an extension method. So if and when a color needs to get contrast it is c.GetContrastColor

I will have a look at that...