@pmiossec 2023-04-10
> The other point. I wanted you to keep it an extension method. So if and when a color needs to get contrast it is c.GetContrastColor

Done.

I have updated the PR to merge into the new `release/4.1` branch also.
If we decide to not include in the 4.1, I will change it back....